# docs/swagger.yaml
basePath: /
definitions:
  handler.Response:
    properties:
      message:
        type: string
    type: object
  models.Product:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      ownerId:
        type: integer
      price:
        type: integer
      quantity:
        type: integer
    type: object
  models.ProductRequest:
    properties:
      description:
        type: string
      name:
        type: string
      price:
        type: integer
      quantity:
        type: integer
    required:
    - description
    - name
    - price
    - quantity
    type: object
  models.Purchase:
    properties:
      cost:
        type: integer
      id:
        type: integer
      productId:
        type: integer
      quantity:
        type: integer
      timestamp:
        type: string
      userId:
        type: integer
    type: object
  models.SignInRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.SignUpRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  models.User:
    properties:
      balance:
        type: integer
      email:
        type: string
      id:
        type: integer
      productList:
        items:
          type: integer
        type: array
      username:
        type: string
    type: object
host: localhost:1334
info:
  contact: {}
  description: API Server for Market Application
  title: Market App API
  version: "1.0"
paths:
  /api/products/:
    get:
      consumes:
      - application/json
      description: List Products
      operationId: list-products
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: ListProducts
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Add Product
      operationId: add-product
      parameters:
      - description: Product Info
        in: body
        name: Input
        required: true
        schema:
          $ref: '#/definitions/models.ProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: AddProduct
      tags:
      - products
  /api/products/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Product
      operationId: delete-product
      parameters:
      - description: ProductID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: DeleteProduct
      tags:
      - products
    get:
      consumes:
      - application/json
      description: Get Product
      operationId: get-product
      parameters:
      - description: ProductID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: GetProduct
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update Product
      operationId: update-product
      parameters:
      - description: Product Info
        in: body
        name: Input
        required: true
        schema:
          $ref: '#/definitions/models.ProductRequest'
      - description: ProductID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: UpdateProduct
      tags:
      - products
  /api/profile:
    get:
      consumes:
      - application/json
      description: Get Your Profile
      operationId: get-profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: GetProfile
      tags:
      - profile
  /api/profile/delete:
    delete:
      consumes:
      - application/json
      description: Delete Profile
      operationId: delete-profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: DeleteProfile
      tags:
      - profile
  /api/profile/deposit:
    post:
      consumes:
      - application/json
      description: Deposit
      operationId: deposit
      parameters:
      - description: Amount
        in: query
        name: amount
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Deposit
      tags:
      - profile
  /api/purchases/{id}:
    post:
      consumes:
      - application/json
      description: Make Purchase
      operationId: make-purchase
      parameters:
      - description: Product Quantity
        in: query
        name: quantity
        type: integer
      - description: ProductID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: MakePurchase
      tags:
      - purchases
  /api/purchases/product/{id}:
    get:
      consumes:
      - application/json
      description: Get Product Purchases
      operationId: get-product-purchases
      parameters:
      - description: ProductID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Purchase'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: GetProductPurchases
      tags:
      - purchases
  /api/purchases/user/{id}:
    get:
      consumes:
      - application/json
      description: Get User Purchases
      operationId: get-user-purchases
      parameters:
      - description: UserID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Purchase'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: GetUserPurchases
      tags:
      - purchases
  /api/users/:
    get:
      consumes:
      - application/json
      description: List Users
      operationId: list-users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: ListUsers
      tags:
      - users
  /api/users/{id}:
    get:
      consumes:
      - application/json
      description: Get User By ID
      operationId: get-user-by-ID
      parameters:
      - description: UserID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: GetUserByID
      tags:
      - users
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Sign Into Account
      operationId: sign-in
      parameters:
      - description: Sign In Info
        in: body
        name: Input
        required: true
        schema:
          $ref: '#/definitions/models.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.Response'
      summary: SignIn
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: Create New Account
      operationId: sign-up
      parameters:
      - description: Sign Up Info
        in: body
        name: Input
        required: true
        schema:
          $ref: '#/definitions/models.SignUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.Response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.Response'
      summary: SignUp
      tags:
      - auth
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"